refactor(frontend): decouple signature status logic from PDF color palette
